MODEL_API/requirements.txt
Django==4.2.7
django-cors-headers==4.2.0
djangorestframework==3.14.0
joblib==1.3.2
scikit-learn==1.3.0